backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
sentence-transformers==2.5.1
qdrant-client==1.7.3
groq==0.4.2
torch==2.2.0
numpy==1.26.4
